refactor: Дистанция заказа в Integer.

+++ app/src/main/java/com/cargopull/executor_driver/backend/websocket/incoming/ApiOrder.java
@@ -17,7 +17,7 @@ public class ApiOrder {
   @SerializedName("overPackageStartCalculationTime")
   private long estimatedTime;
   @SerializedName("estimatedRouteDistance")
-  private long estimatedRouteDistance;
+  private int estimatedRouteDistance;
   @SerializedName("totalAmount")
   private long totalAmount;
   @Nullable
@@ -53,7 +53,7 @@ public class ApiOrder {
 
   @SuppressWarnings("SameParameterValue")
   ApiOrder(long id, @Nullable String estimatedAmountText, long estimatedAmount, long estimatedTime,
-      long estimatedRouteDistance, long totalAmount, @Nullable String comment, long timeout,
+      int estimatedRouteDistance, long totalAmount, @Nullable String comment, long timeout,
       long etaToStartPoint, long confirmationTime, long orderStartTime,
       @Nullable ApiDriverDistancePair executorDistance,
       @Nullable ApiOrderService apiOrderService,
@@ -93,7 +93,7 @@ public class ApiOrder {
     return estimatedTime;
   }
 
-  public long getEstimatedRouteDistance() {
+  public int getEstimatedRouteDistance() {
     return estimatedRouteDistance;
   }
